requirements.txt
# Coinbase Advanced Trade API
coinbase-advanced-py>=1.8.2

# Environment variable management
python-dotenv>=1.0.0

# Additional dependencies for trading bot
requests>=2.31.0
cryptography>=42.0.4
PyJWT>=2.8.0
websockets>=12.0,<14.0
backoff>=2.2.1

# Optional: For data analysis and ML (if needed later)
# pandas>=1.5.0
# numpy>=1.24.0
# scikit-learn>=1.3.0
